requirements.txt
# Core ML (CUDA-enabled)
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4

# Image I/O
Pillow==10.3.0
opencv-python==4.9.0.80

# Data engineering
pyarrow==15.0.2

# Parallel preprocessing
ray==2.10.0

# Experiment tracking
tensorboard==2.16.2

# Configuration
pyyaml==6.0.1

# Testing
pytest==8.1.1